server/requirements.txt
argon2-cffi==20.1.0
cffi==1.14.5
greenlet==1.0.0
grpcio==1.36.1
grpcio-tools==1.36.1
protobuf==3.15.6
psycopg2-binary==2.8.6
pycparser==2.20
six==1.15.0
SQLAlchemy==1.4.4
